Tweaks made to fix highlighting anomalies.
Classified "." as a delimiter, but with no opening (just a closing), so
the strange stuff that happens when you have a period in a string no
longer happens.

Classified "," as a prefix and suffix to numbers so that both sides of
the comma in a statement like HEADING(90,90) are highlighted. This also
resolved other issues with string highlighting (a bit of a nightmare).