A Cheap Workaround
The last commit made it so that keywords that had a period right after
them were not highlighted, so I added 3 new words for on, off and stage,
but with periods at the end. This is unfortunate because now the period
is highlighted too, I hope there is a better way to solve this problem.